Introduce Export::Fetch to build the scope
