+++ lib/export.rb
@@ -3,6 +3,7 @@ require 'active_support/inflector'
 require 'active_record_union'
 require 'active_record'
 require 'export/transform'
+require 'export/fetch'
 require 'export/dependency_tree'
 require 'export/dump'
 require 'export/transform_data'
